Updated invoke-psake function to process the "Parameters" parameter before the "Properties" function of the build script is processed.  This allows you to reference the parameters from within the "Properties" function.
Added new parameter to the invoke-psake function called "Properties" that allows you to override the values of the properties defined in your build script.

Added new examples and specs for using the new invoke-psake parameters "Parameters" and "Properties".

Updated psake-buildTester.ps1 to always pass a standard set of properties so that the spec will pass.

Added code to log the current script-level variables when an exception occurs - this way you have more information to help trouble-shoot errors.