Added -structuredDocs option to Invoke-psake.
The -structuredDocs can be used to return raw (non-formatted) meta data to
the calling script. This makes it possible to consume this meta data from
a script and do something meaningfull with it.

At the same time, some errors were fixed in the documentation (-detailedDocs
was not visible in the help of invoke-psake).

Tests were also added for -docs and -detailedDocs.