(GH185) Prefer default build called psakefile.ps1 over default.ps1
The default psake build file of default.ps1 is completely ambiguous without
inspecting its contents. Rake has a Rakefile, bundle has a Gemfile etc. so
instead, psake should use Psakefile.ps1 as the default build file name.

This commit modifies the load behaviour for a default build file:

- Refactored detecting the default build filename into it's own function

- The `Get-DefaultBuildFile` will prefer to load the `buildFileName` file.  If
it does not exist it fall back to `legacyBuildFileName` and emit a deprecation
warning if it exists.  This allows all existing users to continue to use
`default.ps1` and migrate to psakefile.ps1 when appropriate.