(GH185) Use more descriptive default build file name of psakefile.ps1 (#196)
* (GH185) Prefer default build called psakefile.ps1 over default.ps1

The default psake build file of default.ps1 is completely ambiguous without
inspecting its contents. Rake has a Rakefile, bundle has a Gemfile etc. so
instead, psake should use Psakefile.ps1 as the default build file name.

* Update Framework test for above 4.0

Previously the spec test for the Framework command expected that .Net 4.0 output
a 4.0 version for MSBuild, however, Framework 4.0 will report the latest version
of MSBuild e.g. 4.6.1.  This commit changes the assertion to expect a regular
expression which contains the major version of 4.

* (GH185) Add tests for legacy and default build files

This commit adds test fixtures and spec tests for the default build file loading
behaviour.

* Update build tester to cope with a single test file

Previously if there was only one test file the build tester would fail as it was
not an array type.  This commit detects if the return type is not an array and
if so, forces the return type by a simple array literal.

This commit modifies the load behaviour for a default build file:

- Refactored detecting the default build filename into it's own function

- The `Get-DefaultBuildFile` will prefer to load the `buildFileName` file.  If
it does not exist it fall back to `legacyBuildFileName` and emit a deprecation
warning if it exists.  This allows all existing users to continue to use
`default.ps1` and migrate to psakefile.ps1 when appropriate.

* (GH185) Update documentation with new default build file name

This commit updates existing documentation with the new build file name and adds
a page describing the changes and the rationale behind them.