GH Issue #268 Version Check Too Strict
When checking the version of the from module in task.ps1, the version
check is very strict. A difference in the z release of the installed
module will cause the module to be rejected and the task to fail.

This change adds additional flexibility in the version criteria the user
can specify. The new parameters are `minimumVersion` `maximumVersion`
and `lessThanVersion`. The old `Version` parameter is still present, but
it is now effectively an alias that sets `minimumVersion` and
`maximumVersion` to the same value.

Any of these parameters can be used alone or together to form open ended
or closed ranges of module versions that will acceptable for import.