Remove additional call stack pop from Invoke-Task
The current task is popped from the call stack at the end of the function. When the task fails and -ContinueOnError is specified, the catch block unnecessarily pops the task from the call stack at this point. This causes the pop at the end of the function to actually pop the parent task from the stack.