Ensure Invoke-Task always pops the task from the call stack
This is to resolve a specific scenario where a task with -ContinueOnError depends on a task which fails. The task which fails never pops itself from the stack (it fails before it gets to that point), then the task which depends on in attempts to pop itself from the stack, but actually pops the failed task, which results in a validation error.

To fix this, I've added a try block immediately after the task is pushed onto the stack, and put the code which pops it into a finally block.