Ability to override how psake's internal logging is handled (#301)
<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
There is a new WriteOutput function, which is used throughout the psake
code whenever messages need to be logged. The function simply passes the
message and message type to an "outputHandler" script block, which can
be overridden in psake-config. So you can override exactly how output is
handled by overriding that outputHandler script block.

The default outputHandler script block further routes the output
according to it's output type (warning, verbose, error, etc) to
corresponding script blocks which can also be overridden by
psake-config. So you also have the option of overriding how a single
type of output is handled.

The default outputHandlers for each message type outputs the message
exactly as it would have been prior to this new feature, so it is a
non-breaking change.

## Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->
#65 

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
As described in the issue, if your psake build file uses a different
method of logging to psake's internal logging method, it can be
difficult to try and handle the messages from each consistently. This
change allows you to make psake's internal logging more consistent with
your own. You can even have psake's messages routed to your own custom
functions.

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
New spec files added to verify the basic functionality, that all of
psake's internal logging uses the new WriteOutput function, and that it
handles situations where the config items have been overridden with
invalid values.

## Screenshots (if appropriate):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.

---------

Co-authored-by: Gilbert Sanchez <me@gilbertsanchez.com>